Add files via upload
-----------Cricket Game----------
Welcome to the Cricket Game! This interactive game allows you to make choices as a player and compete against the computer in a simulated cricket match. The game offers three options: "Bat," "Ball," and "Stumps."

How to Play:
Bat: Click the "Bat" button to make your batting choice.
Ball: Click the "Ball" button to make your bowling choice.
Stumps: Click the "Stumps" button to make your stumps choice.

Game Logic:
The computer randomly selects one of the three options (bat, ball, or stumps) after each button click.
The game then determines the winner based on the choices made by the player and the computer.
The result is displayed through an alert, indicating whether you won, the computer won, or it's a tie.

Scenarios:
If you choose "Bat" and the computer chooses "Ball," you win.
If you choose "Ball" and the computer chooses "Stumps," you win.
If you choose "Stumps" and the computer chooses "Bat," you win.
If both you and the computer make the same choice, it's a tie.